// car-gal/interfaces.ts
export interface Config {
    document?: Document | any;
    window?: Window | any;
    rootElement?: HTMLElement | string | null,
    autoInit?: boolean;
    //enableFullScreen?: boolean;
    containerElement?: HTMLElement[] | HTMLElement | string[] | string | null,
    //media?: HTMLElement[] | HTMLElement | Media | Media[];
    defaultOptions?: Options;

    instances?: GalleryInstance[];
    Events?: ConfigEvents;
}

export interface GalleryInstance {
    instance?: IGallery;
    containerId?: string;
    container?: HTMLElement | string | null;
    options: Options;
    media?: HTMLElement[] | IMedia[];
    externalMedia?: HTMLElement[] | IMedia[];
}

export interface ConfigEvents {
    onLoaded?: () => void;
}

export interface Options {
    id?: string;
    enableFullScreen?: boolean;
    //autoInitiate?: boolean;

    // autoplay?: boolean;
    // autoplay_repeat?: boolean;
    // slideInterval?: number;
    lazyLoad?: boolean;

    //thumbnails?: boolean;

    Carousel?: CarouselOptions;
    Fullscreen?: fullscreenOptions;
}

export interface Media {
    type: string;
    src: string;
    srcset: string;
}

export interface CarouselOptions {
    autoplay?: boolean;
    autoplayRepeat?: boolean;
    slideInterval?: number;

    thumbnails?: boolean;

    padding?: string;
    backgroundColor?: string;
    overlayBackground?: string;
    color?: string;

    btns?: btnOptions;
    // btnColor?: string;
    // btnColorHover?: string;
    // btnBackground?: string;
    // btnBackgroundHover?: string;

    //opacity?: number;

    Events?: CarouselEvents;
}

export interface btnOptions {
    color?: string;
    hover?: string;
    background?: string;
    backgroundHover?: string;
}

export interface fullscreenOptions {
    closeOnClick?: boolean;
    background?: string;
    color?: string;
    opacity?: number;
    //menuBarFixed?: boolean;
    Events?: FullscreenEvents;
    Carousel?: CarouselOptions;
    btns?: btnOptions;
    //text?: textOptions;
    Menubar?: menubarOptions;
    title?: IElementCssOptions;
    description?: IElementCssOptions;
}
// ...
export interface IGallery {
    id: number;
    options?: Options;
    containerId?: string;
    container: HTMLElement;
    media: IMedia[];
    externalMedia: IMedia[];
    Carousel?: Carousel;
    Fullscreen?: Fullscreen;
}

export interface IMedia {
    handler?: any;
    origin?: HTMLElement;
    element?: HTMLElement;
    containerElement: HTMLElement;
    type?: string;
    title?: string;
    description?: string;
    sizes?: ISrcSet[];
    currentSizeIndex?: number;
}